Complete day 15
After Srirag confirmed my point was correct I went and looked at the multiplication.
It was overflowing. The reason I didn't pick up on this is because I was running the tests in release
mode, where integer overflow silently wraps instead of panicking.
Why in release mode? For some reason, in debug mode it was unbearably slow (more than 5 seconds).